requirements.txt
langchain
langchain-openai
streamlit
python-dotenv
langchain-community
